Modify redux store
