+++ app/src/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
+import { Provider } from 'react-redux';
+import store from './redux/store';
 import './App.css';
 import Navigation from './components/Navigation';
 import Header from './components/Header';
 import MediaObject from './components/MediaObject';
 import Details from './components/Details';
 import Footer from './components/Footer';
-import store from './store/store';
-import { Provider } from 'react-redux'
 
 function App() {
   return (
