+++ app/src/redux/store.js
@@ -1,6 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
 import campaignReducer from  './campaigns/campaignSlice';
-// import campaignSlice from "./campaigns/campaignSlice";
 
 const store = configureStore({
     reducer: {
